+++ fetcher.go
@@ -64,7 +64,7 @@ func (b *OGame) getPage(page string, celestialID CelestialID, opts ...Option) ([
 		vals = url.Values{"page": {page}}
 	}
 	if celestialID != 0 {
-		vals.Add("cp", FI64(int64(celestialID)))
+		vals.Add("cp", FI64(celestialID))
 	}
 	return b.getPageContent(vals, opts...)
 }
