+++ planetid.go
@@ -4,7 +4,7 @@ package ogame
 type PlanetID CelestialID
 
 func (p PlanetID) String() string {
-	return FI64(int64(p))
+	return FI64(p)
 }
 
 // Celestial convert a PlanetID to a CelestialID
